+++ crates/adam-core/src/time.rs
@@ -1,6 +1,8 @@
 use core::fmt;
 
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd)]
+#[derive(
+    Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd, serde::Serialize, serde::Deserialize,
+)]
 pub struct SimDate {
     year: i32,
     day_of_year: u16,